chore: wip move things around, cleanup, prepare composite types completion
